requirements.txt
Appium-Python-Client==3.1.1
pytest==7.4.0